Streamlitapp/Grade-07/requirements.txt
numpy==1.20.1
Pillow==9.3.0
PyAutoGUI==0.9.53
scipy==1.6.2
streamlit==1.15.2
streamlit_option_menu==0.3.2

